requirements.txt
annotated-types==0.7.0
anthropic==0.34.1
anyio==4.4.0
boto3==1.35.11
botocore==1.35.11
cachetools==5.5.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.143.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
groq==0.10.0
grpcio==1.66.1
grpcio-status==1.62.3
guidance==0.1.16
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsonpath-python==1.0.6
kiwisolver==1.4.7
llama_cpp_python==0.2.89
MarkupSafe==2.1.5
matplotlib==3.9.2
mistralai
mpmath==1.3.0
mypy-extensions==1.0.0
networkx
numpy
openai==1.42.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
proto-plus==1.24.0
protobuf==4.25.4
py==1.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
retry==0.9.2
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.4
seaborn==0.13.2
setuptools==74.0.0
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3
